+++ RSS-Services/DTOs/PlayerStatsDTO.cs
@@ -3,8 +3,8 @@ namespace RSS_Services.DTOs
     public class PlayerStatsDTO
     {
         public int PeriodsWon { get; set; }
-        public int TotalWagered { get; set; }
-        public int WagersWon { get; set; }
+        public decimal TotalWagered { get; set; }
+        public decimal WagersWon { get; set; }
         public double WinRate { get; set; }
         public int TotalSquaresClaimed { get; set; }
     }
